Refactor collision stuff into the correct package. Refactor some math util. Adjust how jumping respects gravity
