+++ jump-core/src/main/java/com/bitdecay/jump/collision/BitResolution.java
@@ -1,5 +1,6 @@
-package com.bitdecay.jump;
+package com.bitdecay.jump.collision;
 
+import com.bitdecay.jump.BitBody;
 import com.bitdecay.jump.geom.BitPoint;
 import com.bitdecay.jump.geom.BitRectangle;
 import java.util.PriorityQueue;
@@ -16,12 +17,14 @@ public abstract class BitResolution {
 	 * A priority queue that resolves the LARGEST collisions first. Larger
 	 * collisions (by area) are easier to determine accurate resolutions for.
 	 */
-	protected PriorityQueue<BitCollision> collisions = new PriorityQueue<BitCollision>();
+	public PriorityQueue<BitCollision> collisions = new PriorityQueue<BitCollision>();
+
 	protected BitRectangle resolvedPosition;
 	protected BitPoint resolution = new BitPoint(0, 0);
-	protected boolean haltX;
-	protected boolean haltY;
 	protected BitBody body;
+
+	public boolean haltX;
+	public boolean haltY;
 
 	public BitResolution(BitBody body) {
 		this.body = body;
@@ -47,4 +50,12 @@ public abstract class BitResolution {
 	 * squeezed between two bodies)
 	 */
 	public abstract void satisfy(BitWorld world);
+
+	public boolean haltX() {
+		return haltX;
+	}
+
+	public boolean haltY() {
+		return haltY;
+	}
 }
