+++ jump-core/src/main/java/com/bitdecay/jump/collision/BitWorld.java
@@ -1,9 +1,9 @@
-package com.bitdecay.jump;
+package com.bitdecay.jump.collision;
 
 import java.util.*;
 
-import com.bitdecay.jump.collision.ContactListener;
-import com.bitdecay.jump.collision.SATStrategy;
+import com.bitdecay.jump.BitBody;
+import com.bitdecay.jump.BodyType;
 import com.bitdecay.jump.geom.*;
 import com.bitdecay.jump.level.Level;
 import com.bitdecay.jump.level.TileObject;
@@ -355,20 +355,13 @@ public class BitWorld {
 	private void applyResolution(BitBody body, BitResolution resolution) {
 		if (resolution.resolution.x != 0 || resolution.resolution.y != 0) {
 			body.aabb.translate(resolution.resolution);
-			// CONSIDER: have grounded check based on gravity direction rather than just always assuming down
-			if (Math.abs(gravity.y - resolution.resolution.y) > Math.abs(gravity.y)) {
-				// if the body was resolved against the gravity's y, we assume grounded.
-				// CONSIDER: 4-directional gravity might become a possibility.
-				body.grounded = true;
+			if (resolution.haltX) {
+				body.velocity.x = 0;
+			}
+			if (resolution.haltY) {
+				body.velocity.y = 0;
 			}
 		}
-		if (resolution.haltX) {
-			body.velocity.x = 0;
-		}
-		if (resolution.haltY) {
-			body.velocity.y = 0;
-		}
-
 		body.lastResolution = resolution.resolution;
 	}
