+++ jump-leveleditor/src/main/java/com/bitdecay/jump/leveleditor/entity/Player.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g2d.Animation;
 import com.badlogic.gdx.graphics.g2d.Animation.PlayMode;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.bitdecay.jump.BitBody;
-import com.bitdecay.jump.BitWorld;
+import com.bitdecay.jump.collision.BitWorld;
 import com.bitdecay.jump.JumperBody;
 import com.bitdecay.jump.gdx.engine.animation.AnimationState;
 import com.bitdecay.jump.gdx.engine.utilities.ImageUtilities;
